Use vertex templates to handle float/fp16 datatypes
Reviewers: richard

Reviewed By: richard

Differential Revision: https://phabricator.sourcevertex.net/D309

+++ Layer.cpp
@@ -19,7 +19,7 @@ const std::string &Layer::getDType() const { return net.getDType(); }
 unsigned Layer::getDTypeSize() const {
   if (getDType() == "float")
     return 4;
-  assert(getDType() == "short");
+  assert(getDType() == "half");
   return 2;
 }
